FIX: folder deletion bug

+++ app/Http/Controllers/Api/FoldersController.php
@@ -123,12 +123,21 @@ class FoldersController extends Controller
     public function delete($year, $semester, $folder_id)
     {
         $folder = Folder::find($folder_id);
+
+        if (!$folder) {
+            return response([
+                "status" => 0,
+                "message" => "No se encontro la carpeta",
+            ], 404);
+        }
+
         $standardId = $folder->standard_id;
         $typeEvidenceId = $folder->evidence_type_id;
         $currentPath = 'evidencias/' . $year . '/' . $semester . '/' .'estandar_' . $standardId . '/tipo_evidencia_'. $typeEvidenceId;
         $currentFolderPath = $currentPath . $folder->path;
         // Obtiene las evidencias vinculadas y las borra
         $evidences = Evidence::where('folder_id', $folder_id)->get();
+
         if ($evidences) {
             foreach ($evidences as $evidence) {
                 $evidence->delete();
